add a Readline class (and a test!)
Readline.dispatch behaves just like Basil.dispatch except it handles
command substitution and pipe-lining. 00-pipe.rb and the echo command
have been prototyping these ideas, but I think it should be handled
within basil.

Enabling this will be a Config choice (assuming it works), and I've got
plans to make a plugin for runtime switching of whitelisted configs so
it can be further played with in chat.